pyproject.toml
[project]
name = "weather-discord"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "npkk", email = "koba.dev.npkk@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "pydantic>=2.12.3",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.44",
]

[project.scripts]
weather-discord = "weather_discord:main"

[build-system]
requires = ["uv_build>=0.9.5,<0.10.0"]
build-backend = "uv_build"
